+++ projects/ng-gallery-demo/src/app/pages/gallerize-example/gallerize-example.component.ts
@@ -1,7 +1,15 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { NgFor, AsyncPipe, SlicePipe } from '@angular/common';
 import { Title } from '@angular/platform-browser';
-import { Gallery } from 'ng-gallery';
+import { RouterLink } from '@angular/router';
+import { MatButtonModule } from '@angular/material/button';
+import { Gallery, GalleryComponent } from 'ng-gallery';
+import { GallerizeDirective } from 'ng-gallery/lightbox';
 import { Pixabay } from '../../service/pixabay.service';
+import { FooterComponent } from '../../shared/footer/footer.component';
+import { HlCodeComponent } from '../../shared/hl-code/hl-code.component';
+import { NoteComponent } from '../../shared/note/note.component';
+import { SectionTitleComponent } from '../../shared/section-title/section-title.component';
 
 @Component({
   host: {
@@ -10,7 +18,9 @@ import { Pixabay } from '../../service/pixabay.service';
   selector: 'gallerize-example',
   templateUrl: './gallerize-example.component.html',
   styleUrls: ['./gallerize-example.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [SectionTitleComponent, NoteComponent, MatButtonModule, RouterLink, HlCodeComponent, GallerizeDirective, NgFor, GalleryComponent, FooterComponent, AsyncPipe, SlicePipe]
 })
 export class GallerizeExampleComponent implements OnInit {
 
@@ -25,6 +35,7 @@ export class GallerizeExampleComponent implements OnInit {
       imageSize: 'cover',
       itemAutosize: false,
       thumbAutosize: false,
+      thumbs: true
       // thumbWidth: 120,
       // thumbHeight: 90,
       // thumbView: 'contain'
@@ -35,6 +46,7 @@ export class GallerizeExampleComponent implements OnInit {
       imageSize: 'cover',
       itemAutosize: false,
       thumbAutosize: false,
+      thumbs: true
       // thumbView: 'contain'
     });
   }
@@ -48,9 +60,9 @@ export class GallerizeExampleComponent implements OnInit {
 const code = {
   onElements: `<div class="container" gallerize>
   <img *ngFor="let image of images"
-       src="{{image.thumb}}"
+       [src]="image.thumb"
        [attr.imageSrc]="image.src"
-       [attr.thumbSrc]="image.thumb">
+       [attr.thumbSrc]="image.thumb"/>
 </div>`,
   withSelector: `<div class="container" gallerize selector=".gallery-img">
   <div *ngFor="let image of images" class="gallery-img"
@@ -58,5 +70,5 @@ const code = {
        [attr.imageSrc]="image.src"
        [attr.thumbSrc]="image.thumb"></div>
 </div>`,
-  onGallery: `<gallery gallerize [items]="cameraImages"></gallery>`
+  onGallery: `<gallery thumbs gallerize [items]="cameraImages"></gallery>`
 };