+++ projects/ng-gallery-demo/src/app/routing.module.ts
@@ -1,50 +1,49 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { Routes } from '@angular/router';
 
-const appRoutes: Routes = [
+export const appRoutes: Routes = [
   {
     path: '',
-    loadChildren: () => import('./pages/home/home.module').then(m => m.HomeModule),
+    loadComponent: () => import('./pages/home/home.component').then(m => m.HomeComponent),
     pathMatch: 'full'
   },
   {
     path: 'getting-started',
-    loadChildren: () => import('./pages/documentation/documentation.module').then(m => m.DocumentationModule)
+    loadChildren: () => import('./pages/documentation/routes').then(m => m.DOCUMENTATION_ROUTES)
   },
   {
     path: 'gallery',
-    loadChildren: () => import('./pages/gallery-example/gallery-example.module').then(m => m.GalleryExampleModule)
+    loadComponent: () => import('./pages/gallery-example/gallery-example.component').then(m => m.GalleryExampleComponent)
   },
   {
     path: 'lightbox',
-    loadChildren: () => import('./pages/lightbox-example/lightbox-example.module').then(m => m.LightboxExampleModule)
+    loadComponent: () => import('./pages/lightbox-example/lightbox-example.component').then(m => m.LightboxExampleComponent)
   },
   {
     path: 'gallerize',
-    loadChildren: () => import('./pages/gallerize-example/gallerize-example.module').then(m => m.GallerizeExampleModule)
+    loadComponent: () => import('./pages/gallerize-example/gallerize-example.component').then(m => m.GallerizeExampleComponent)
   },
   {
-    path: 'advanced',
-    loadChildren: () => import('./pages/advanced-example/advanced-example.module').then(m => m.AdvancedExampleModule)
+    path: 'custom-templates',
+    loadComponent: () => import('./pages/templates-example/templates-example.component').then(m => m.TemplatesExampleComponent)
   },
   {
     path: 'lab',
-    loadChildren: () => import('./pages/lab/lab.module').then(m => m.LabModule)
+    loadComponent: () => import('./pages/lab/lab.component').then(m => m.LabComponent)
   },
   {
     path: '**',
-    loadChildren: () => import('./pages/not-found/not-found.module').then(m => m.NotFoundPageModule)
+    loadComponent: () => import('./pages/not-found/not-found.component').then(m => m.NotFoundComponent)
   }
 ];
 
-@NgModule({
-  imports: [
-    RouterModule.forRoot(appRoutes, {useHash: true})
-  ],
-  exports: [
-    RouterModule
-  ]
-})
-
-export class AppRoutingModule {
-}
+// @NgModule({
+//   imports: [
+//     RouterModule.forRoot(appRoutes, {useHash: true})
+//   ],
+//   exports: [
+//     RouterModule
+//   ]
+// })
+//
+// export class AppRoutingModule {
+// }