+++ projects/ng-gallery-demo/src/app/service/pixabay.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable, shareReplay } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Observable, shareReplay, map } from 'rxjs';
 
 import { ImageItem, GalleryItem } from 'ng-gallery';
 
@@ -12,7 +11,7 @@ import { Hit2, PixabayHDModel } from './pixabay.model';
 })
 export class Pixabay {
 
-  private readonly API_KEY = '560162-704dd2880c027f22c62ab7941';
+  private readonly API_KEY: string = '560162-704dd2880c027f22c62ab7941';
 
   constructor(private _http: HttpClient) {
   }