+++ projects/ng-gallery-demo/src/app/shared/hl-code/hl-code.component.ts
@@ -1,20 +1,27 @@
-import { Component, Input, ChangeDetectionStrategy } from '@angular/core';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { Component, Input, inject, ChangeDetectionStrategy, booleanAttribute } from '@angular/core';
+import { ClipboardModule } from '@angular/cdk/clipboard';
+import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatTooltip } from '@angular/material/tooltip';
+import { NgScrollbar } from 'ngx-scrollbar';
+import { HighlightAuto } from 'ngx-highlightjs';
 
 @Component({
+  standalone: true,
   selector: 'hl-code',
   templateUrl: './hl-code.component.html',
   styleUrls: ['./hl-code.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [MatSnackBarModule, NgScrollbar, MatButtonModule, ClipboardModule, MatIconModule, HighlightAuto, MatTooltip]
 })
 export class HlCodeComponent {
 
-  @Input() code: string;
-  @Input() disabled: boolean;
-  @Input() track = 'vertical';
+  readonly snackBar: MatSnackBar = inject(MatSnackBar);
 
-  constructor(public snackBar: MatSnackBar) {
-  }
+  @Input() code: string;
+  @Input() track: string = 'vertical';
+  @Input({ transform: booleanAttribute }) disabled: boolean;
 
 }