+++ projects/ng-gallery/src/lib/components/gallery-thumb.component.ts
@@ -1,9 +1,23 @@
-import { Component, Input, ChangeDetectionStrategy, HostBinding, Output, EventEmitter } from '@angular/core';
+import {
+  Component,
+  Input,
+  Output,
+  HostBinding,
+  EventEmitter,
+  ElementRef,
+  ChangeDetectionStrategy
+} from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { GalleryItemContext } from '../directives/gallery-item-def.directive';
+import { GalleryImageComponent } from './templates/gallery-image.component';
+import { ImageItemData } from './templates/items.model';
 import { GalleryConfig } from '../models/config.model';
+import { GalleryItemType } from '../models/constants';
 
 @Component({
   selector: 'gallery-thumb',
   changeDetection: ChangeDetectionStrategy.OnPush,
+  styleUrls: ['./gallery-thumb.scss'],
   template: `
     <gallery-image [src]="data.thumb"
                    [alt]="data.alt + '-thumbnail'"
@@ -13,11 +27,11 @@ import { GalleryConfig } from '../models/config.model';
                    (error)="error.emit($event)"></gallery-image>
 
     <div *ngIf="config.thumbTemplate" class="g-template g-thumb-template">
-      <ng-container
-        *ngTemplateOutlet="config.thumbTemplate; context: { index, type, data, isActive }">
-      </ng-container>
+      <ng-container *ngTemplateOutlet="config.thumbTemplate; context: imageContext"></ng-container>
     </div>
-  `
+  `,
+  standalone: true,
+  imports: [CommonModule, GalleryImageComponent]
 })
 export class GalleryThumbComponent {
 
@@ -26,19 +40,44 @@ export class GalleryThumbComponent {
   /** Item's index in the gallery */
   @Input() index: number;
 
+  /** The number of total items */
+  @Input() count: number;
+
   /** Gallery current index */
   @Input() currIndex: number;
 
   /** Item's type 'image', 'video', 'youtube', 'iframe' */
-  @Input() type: string;
+  @Input() type: GalleryItemType;
 
   /** Item's data, this object contains the data required to display the content (e.g. src path) */
-  @Input() data: any;
+  @Input() data: ImageItemData;
 
-  @Output() error = new EventEmitter<ErrorEvent>();
+  @Output() error: EventEmitter<ErrorEvent> = new EventEmitter<ErrorEvent>();
 
   @HostBinding('class.g-active-thumb') get isActive() {
     return this.index === this.currIndex;
   }
 
+  @HostBinding('attr.galleryIndex') get isIndexAttr(): number {
+    return this.index;
+  }
+
+  get imageContext(): GalleryItemContext<ImageItemData> {
+    return {
+      $implicit: this.data,
+      index: this.index,
+      type: this.type,
+      active: this.isActive,
+      count: this.count,
+      first: this.index === 0,
+      last: this.index === this.count - 1
+    }
+  }
+
+  get nativeElement(): HTMLElement {
+    return this.el.nativeElement;
+  }
+
+  constructor(private el: ElementRef<HTMLElement>) {
+  }
 }