+++ projects/ng-gallery/src/lib/components/templates/gallery-video.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, Output, EventEmitter, OnInit, ViewChild, ElementRef, ChangeDetectionStrategy } from '@angular/core';
+import { NgFor, NgIf } from '@angular/common';
 
 @Component({
   selector: 'gallery-video',
@@ -13,9 +14,16 @@ import { Component, Input, Output, EventEmitter, OnInit, ViewChild, ElementRef, 
            [loop]="loop"
            [poster]="poster"
            (error)="error.emit($event)">
-      <source *ngFor="let src of videoSources" [src]="src?.url" [type]="src?.type"/>
+      <ng-container *ngFor="let src of videoSources">
+        <source *ngIf="src?.type; else noType" [src]="src?.url" [type]="src.type"/>
+        <ng-template #noType>
+          <source [src]="src?.url"/>
+        </ng-template>
+      </ng-container>
     </video>
-  `
+  `,
+  standalone: true,
+  imports: [NgFor, NgIf]
 })
 export class GalleryVideoComponent implements OnInit {