+++ projects/ng-gallery/src/lib/gallery.module.ts
@@ -1,57 +1,25 @@
-import { NgModule, ModuleWithProviders } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { PlatformModule } from '@angular/cdk/platform';
-
-import { GalleryConfig, GALLERY_CONFIG } from './models/config.model';
-
+import { NgModule } from '@angular/core';
 import { GalleryComponent } from './components/gallery.component';
-import { GalleryNavComponent } from './components/gallery-nav.component';
-import { GalleryCoreComponent } from './components/gallery-core.component';
-import { GalleryDotsComponent } from './components/gallery-dots.component';
-import { GalleryThumbsComponent } from './components/gallery-thumbs.component';
-import { GallerySliderComponent } from './components/gallery-slider.component';
-import { GalleryCounterComponent } from './components/gallery-counter.component';
-
-import { GalleryItemComponent } from './components/gallery-item.component';
-import { GalleryThumbComponent } from './components/gallery-thumb.component';
-import { GalleryImageComponent } from './components/templates/gallery-image.component';
-import { GalleryVideoComponent } from './components/templates/gallery-video.component';
-import { GalleryIframeComponent } from './components/templates/gallery-iframe.component';
+import { GalleryImageDef } from './directives/gallery-image-def.directive';
+import { GalleryThumbDef } from './directives/gallery-thumb-def.directive';
+import { GalleryItemDef } from './directives/gallery-item-def.directive';
+import { GalleryBoxDef } from './directives/gallery-box-def.directive';
 
 @NgModule({
   imports: [
-    CommonModule,
-    PlatformModule
-  ],
-  declarations: [
     GalleryComponent,
-    GalleryNavComponent,
-    GalleryDotsComponent,
-    GalleryCoreComponent,
-    GallerySliderComponent,
-    GalleryCounterComponent,
-    GalleryThumbsComponent,
-    GalleryThumbComponent,
-    GalleryItemComponent,
-    GalleryImageComponent,
-    GalleryVideoComponent,
-    GalleryIframeComponent
+    GalleryItemDef,
+    GalleryImageDef,
+    GalleryThumbDef,
+    GalleryBoxDef
   ],
   exports: [
-    GalleryComponent
+    GalleryComponent,
+    GalleryItemDef,
+    GalleryImageDef,
+    GalleryThumbDef,
+    GalleryBoxDef
   ]
 })
 export class GalleryModule {
-  static withConfig(config: GalleryConfig): ModuleWithProviders<GalleryModule> {
-
-    return {
-      ngModule: GalleryModule,
-      providers: [
-        {
-          provide: GALLERY_CONFIG,
-          useValue: config
-        }
-      ]
-    };
-  }
 }